before refactor to immer and lodash
