+++ package.json
@@ -48,6 +48,7 @@
     "@testing-library/user-event": "^12.5.0",
     "@types/electron-devtools-installer": "^2.2.0",
     "@types/jest": "^26.0.3",
+    "@types/lodash": "^4.14.168",
     "@types/react": "^16.9.34",
     "@types/react-dom": "^16.9.6",
     "@types/react-router-dom": "^5.1.6",
@@ -78,7 +79,7 @@
     "style-loader": "^2.0.0",
     "ts-jest": "^26.1.1",
     "typescript": "^3.9.5",
-    "wait-on": "^5.0.1",
+    "wait-on": "^5.3.0",
     "webpack": "^4.42.1",
     "webpack-cli": "^3.3.10",
     "webpack-dev-server": "^3.10.3"
@@ -92,6 +93,8 @@
     "formik": "^2.2.5",
     "formik-material-ui": "^3.0.0",
     "i18next": "^19.8.4",
+    "immer": "^9.0.1",
+    "lodash": "^4.17.21",
     "react": "^16.13.1",
     "react-dom": "^16.13.1",
     "react-i18next": "^11.7.4",
