+++ src/components/dashboard/utils.ts
@@ -1,3 +1,4 @@
+import { TFunction } from 'i18next';
 import {
   HomeResponseTreeChildren,
   HomeFormTreeResponse,
@@ -160,3 +161,61 @@ export const handleExpand = (matches: string[], id: string) => {
   }
   return [...matches, id];
 };
+
+const SMALL_SIZE = 100;
+const MEDIUM_SIZE = 200;
+const LARGE_SIZE = 300;
+
+export const columns = (t: TFunction) => [
+  {
+    field: 'id',
+    headerName: t('dashboard.dashboardTable.header.ID'),
+    flex: SMALL_SIZE,
+  },
+  {
+    field: 'OperandIdentifier',
+    headerName: t(
+      'dashboard.dashboardTable.header.operandIdentifier',
+    ),
+    flex: MEDIUM_SIZE,
+  },
+  {
+    field: 'Address',
+    headerName: t('dashboard.dashboardTable.header.address'),
+    flex: SMALL_SIZE,
+  },
+  {
+    field: 'Name',
+    headerName: t('dashboard.dashboardTable.header.name'),
+    flex: LARGE_SIZE,
+  },
+  {
+    field: 'DataType',
+    headerName: t('dashboard.dashboardTable.header.dataType'),
+    flex: SMALL_SIZE,
+  },
+  {
+    field: 'Comment',
+    headerName: t('dashboard.dashboardTable.header.comment'),
+    flex: LARGE_SIZE,
+  },
+  {
+    field: 'AccessibleFromHmi',
+    headerName: t(
+      'dashboard.dashboardTable.header.accessibleFromHmi',
+    ),
+    flex: MEDIUM_SIZE,
+  },
+  {
+    field: 'VisibleInHmiEngineering',
+    headerName: t(
+      'dashboard.dashboardTable.header.visibleInHmiEngineering',
+    ),
+    flex: MEDIUM_SIZE,
+  },
+  {
+    field: 'WritableFromHmi',
+    headerName: t('dashboard.dashboardTable.header.writableFromHmi'),
+    flex: MEDIUM_SIZE,
+  },
+];
