+++ src/components/home/HomeForm.tsx
@@ -12,15 +12,18 @@ import { makeStyles } from '@material-ui/core/styles';
 import Card from '@material-ui/core/Card';
 import FormControl from '@material-ui/core/FormControl';
 import Button from '@material-ui/core/Button';
-import ListItemText from '@material-ui/core/ListItemText';
-import ListItem from '@material-ui/core/ListItem';
 import List from '@material-ui/core/List';
-import ListItemSecondaryAction from '@material-ui/core/ListItemSecondaryAction';
-import IconButton from '@material-ui/core/IconButton';
-import DeleteIcon from '@material-ui/icons/Delete';
 import Typography from '@material-ui/core/Typography';
 
-import { HomeFormValues, FileItem } from 'types';
+import HomeFileItemList from './HomeFileItemList';
+import {
+  handleFileAppend,
+  handleRemoveItem,
+  initialValues,
+  isDisabled,
+} from './utils';
+
+import { HomeFormValues } from 'types';
 
 type HomeFormProps = {
   handleSubmit: (
@@ -50,47 +53,10 @@ const useStyles = makeStyles(() => ({
   },
 }));
 
-const initialValues: HomeFormValues = {
-  projectName: '',
-  eplanTags: [],
-  spsMatrix: [],
-};
-
-const isDisabled = (values: HomeFormValues) => {
-  return values.eplanTags.length === 0 || values.projectName === '';
-};
-
 const HomeForm: React.FC<HomeFormProps> = (props) => {
   const { handleSubmit } = props;
   const { t } = useTranslation();
   const classes = useStyles();
-
-  const handleChange = (
-    push: (element: FileItem) => void,
-    values: FileItem[],
-  ) => (event: React.ChangeEvent<HTMLInputElement>) => {
-    event.persist();
-    const file =
-      (event.target.files && event.target.files[0]) || null;
-
-    if (file != null) {
-      const element: FileItem = { name: file.name, path: file.path };
-      if (
-        !values.some(
-          (file) => JSON.stringify(file) === JSON.stringify(element),
-        )
-      ) {
-        push(element);
-      }
-    }
-  };
-
-  const handleClick = (
-    e: React.MouseEvent<HTMLInputElement, MouseEvent>,
-  ) => {
-    //@ts-ignore
-    e.target.value = null;
-  };
 
   return (
     <Card className={classes.formContainer}>
@@ -124,11 +90,11 @@ const HomeForm: React.FC<HomeFormProps> = (props) => {
                         id="upload-eplanTags"
                         name="eplanTags"
                         type="file"
-                        onChange={handleChange(
+                        onChange={handleFileAppend(
                           push,
                           values.eplanTags,
                         )}
-                        onClick={handleClick}
+                        onClick={handleRemoveItem}
                         required
                       />
                       <Button
@@ -142,21 +108,13 @@ const HomeForm: React.FC<HomeFormProps> = (props) => {
                     </label>
                   </FormControl>
                   <List dense={true}>
-                    {values.eplanTags.map((el, index) => (
-                      <React.Fragment key={el.name}>
-                        <ListItem disableGutters>
-                          <ListItemText primary={el.name} />
-                          <ListItemSecondaryAction>
-                            <IconButton
-                              edge="end"
-                              aria-label={`file-eplan-${index}`}
-                              onClick={() => remove(index)}
-                            >
-                              <DeleteIcon />
-                            </IconButton>
-                          </ListItemSecondaryAction>
-                        </ListItem>
-                      </React.Fragment>
+                    {values.eplanTags.map((eplan, index) => (
+                      <HomeFileItemList
+                        key={eplan.name}
+                        value={eplan}
+                        label={`file-eplan-${index}`}
+                        onClick={() => remove(index)}
+                      />
                     ))}
                   </List>
                 </>
@@ -176,11 +134,11 @@ const HomeForm: React.FC<HomeFormProps> = (props) => {
                         id="upload-spsMatrix"
                         name="spsMatrix"
                         type="file"
-                        onChange={handleChange(
+                        onChange={handleFileAppend(
                           push,
                           values.spsMatrix,
                         )}
-                        onClick={handleClick}
+                        onClick={handleRemoveItem}
                       />
                       <Button
                         color="secondary"
@@ -193,21 +151,13 @@ const HomeForm: React.FC<HomeFormProps> = (props) => {
                     </label>
                   </FormControl>
                   <List dense={true}>
-                    {values.spsMatrix.map((el, index) => (
-                      <React.Fragment key={el.name}>
-                        <ListItem disableGutters>
-                          <ListItemText primary={el.name} />
-                          <ListItemSecondaryAction>
-                            <IconButton
-                              edge="end"
-                              aria-label={`file-sps-${index}`}
-                              onClick={() => remove(index)}
-                            >
-                              <DeleteIcon />
-                            </IconButton>
-                          </ListItemSecondaryAction>
-                        </ListItem>
-                      </React.Fragment>
+                    {values.spsMatrix.map((spsMatrix, index) => (
+                      <HomeFileItemList
+                        key={spsMatrix.name}
+                        value={spsMatrix}
+                        label={`file-sps-${index}`}
+                        onClick={() => remove(index)}
+                      />
                     ))}
                   </List>
                 </>
