+++ src/i18n/resources/en/components/dashboard/dashboardNavigation.ts
@@ -1,4 +1,5 @@
 export const dashboardNavigationTranslations = {
   'dashboard.navigation.tab.table': 'Unused Symbols',
   'dashboard.navigation.tab.tree': 'Structure',
+  'dashboard.navigation.tab.process': 'Process',
 };
