+++ src/i18n/resources/pl/components/dashboard/dashboardNavigation.ts
@@ -1,4 +1,5 @@
 export const dashboardNavigationTranslations = {
   'dashboard.navigation.tab.table': 'Nieużywane symbole',
   'dashboard.navigation.tab.tree': 'Struktura',
+  'dashboard.navigation.tab.process': 'Proces',
 };
