+++ src/pages/Process.tsx
@@ -1,0 +1,7 @@
+import React from 'react';
+
+const Process = () => {
+  return <div></div>;
+};
+
+export default Process;
