+++ src/utils/objectUtils.ts
@@ -1,0 +1,5 @@
+import _ from 'lodash';
+
+export const isDeepEqual = (value: any, other: any) => {
+  return _.isEqual(value, other);
+};
