+++ tsconfig.json
@@ -2,12 +2,7 @@
   "compilerOptions": {
     "target": "es5",
     "module": "commonjs",
-    "lib": [
-      "dom",
-      "es2015",
-      "es2016",
-      "es2017"
-    ],
+    "lib": ["dom", "es2015", "es2016", "es2017"],
     "allowJs": true,
     "jsx": "react",
     "sourceMap": true,
@@ -16,9 +11,8 @@
     "esModuleInterop": true,
     "baseUrl": "src",
     "paths": {
-      "test-utils": [
-        "./tests/test-utils"
-      ]
+      "test-utils": ["./tests/test-utils"]
     }
-  }
+  },
+  "include": ["src"]
 }
