+++ webpack/electron.webpack.js
@@ -5,6 +5,12 @@ const rootPath = path.resolve(__dirname, '..');
 module.exports = {
   resolve: {
     extensions: ['.tsx', '.ts', '.js'],
+    mainFields: ['main', 'module', 'browser'],
+    modules: [
+      path.resolve(__dirname, 'src'),
+      path.resolve(__dirname, 'tests'),
+      'node_modules',
+    ],
   },
   devtool: 'source-map',
   entry: path.resolve(rootPath, 'electron', 'main.js'),
