+++ webpack/react.webpack.js
@@ -5,11 +5,16 @@ const HtmlWebpackPlugin = require('html-webpack-plugin');
 const rootPath = path.resolve(__dirname, '..');
 const IP = 'http://localhost:5000';
 
+console.log(__dirname);
 module.exports = (env) => ({
   resolve: {
     extensions: ['.tsx', '.ts', '.js'],
     mainFields: ['main', 'module', 'browser'],
-    modules: [path.resolve(__dirname, 'src'), 'node_modules'],
+    modules: [
+      path.resolve(rootPath, 'src'),
+      path.resolve(rootPath, 'tests'),
+      'node_modules',
+    ],
   },
   entry: path.resolve(rootPath, 'src', 'App.tsx'),
   target: 'electron-renderer',
